refactor(tests): split EarmarkStoreTests (925→≤189) (#272, #267)
Split MoolahTests/Features/EarmarkStoreTests.swift into seven
@Suite-per-file files along thematic boundaries (load, applyDelta,
converted balances, reorder, create/update/hide, visibility &
multi-instrument, partial conversion). Every new file sits well under
the file_length 400-line threshold (largest is 189).

The split also eliminates all 22 baselined `identifier_name`
violations (`e0`/`e1`/`e2`) by renaming in the same pass — those names
would otherwise re-key to new file paths. Replacements are
domain-descriptive at the call site (`holiday`/`carRepair`,
`first`/`second`/`third`, `higher`/`lower`,
`firstVisible`/`hidden`/`secondVisible`, `only`).

Also eliminates the old file's `type_body_length` entry (the big type
is now seven smaller types).

Stacked on #351 until it lands.
